refactor(coordinator): split worker supervisor responsibilities
